ai-service/requirements.txt
flask==3.0.2
langchain==0.1.13
langchain-community==0.0.27
langchain-openai==0.0.8
openai==1.12.0
pypdf2==3.0.1
python-dotenv==1.0.1
